fix(agentbox): unblock build by moving cli-snapshot types out of api file
Dockerfile.agentbox does not COPY src/portal/, but PR #237 added type
imports from src/portal/cli-snapshot-api.js into src/core/ — the agentbox
tsc step then fails with "Cannot find module" because the file is not
in the build context. cli-snapshot-api.ts also imports gateway/db.js and
gateway/rest-router.js (Portal runtime code) which are intentionally out
of the agentbox tsconfig include list.

Extract the public type interfaces into src/portal/cli-snapshot-types.ts
(no runtime imports), have cli-snapshot-api.ts re-export them for any
existing importer, and point the three core/ callers at the new file.
Add a single targeted COPY in Dockerfile.agentbox so tsc can resolve the
types without pulling Portal runtime code into the agentbox image.

This fix was discovered while building images to deploy PR #245 to a
test cluster — the bug exists on main since #237 landed.
